+++ spec/services/imports_lms_users/imports_lms_users_spec.rb
@@ -1,4 +1,4 @@
-describe Services::Actions::ImportsLMSUsersById do
+describe Services::Actions::ImportsLMSUsers do
   let(:course) { create :course }
   let(:provider) { :canvas }
   let(:student) { User.unscoped.last }
